utils: traversal utilities structures
